projects/multimodal-game-agent/pyproject.toml
[project]
name = "multimodal-game-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "albumentations>=2.0.6",
    "fastapi[standard]>=0.115.12",
    "line-profiler>=4.2.0",
    "num2words>=0.5.14",
    "openai>=1.76.2",
    "owa",
    "pandas>=2.2.3",
    "pillow>=11.1.0",
    "torch==2.6.0",
    "torchvision==0.21.0",
    "tqdm>=4.67.1",
    "transformers>=4.51.3",
    "trl>=0.16.0",
    "typer>=0.15.2",
    "uvicorn>=0.34.0",
    "wandb>=0.19.10",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/owa_game_agent"]

[tool.uv.sources]
torch = [
  { index = "pytorch-cu126", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
torchvision = [
  { index = "pytorch-cu126", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
owa = { path = "../../" }

[[tool.uv.index]]
name = "pytorch-cu126"
url = "https://download.pytorch.org/whl/cu126"
explicit = true
